Rename to `MultiMQTTRoomService`
