+++ src/components/NetworkRequestIndicator/NetworkRequestIndicator.tsx
@@ -3,7 +3,7 @@ import React, { useMemo } from "react";
 import SyncIcon from "@mui/icons-material/Sync";
 import IconButton from "@mui/material/IconButton";
 
-import { useMultiMQTTRoomContext } from "@services/MQTTRoomService/react";
+import { useMultiMQTTRoomContext } from "@services/MultiMQTTRoomService/react";
 import clsx from "clsx";
 import { useNavigate } from "react-router-dom";
