+++ src/components/SharedSessionManager/SharedSessionManager.ConnectForm.tsx
@@ -3,7 +3,7 @@ import React, { useCallback, useState } from "react";
 import { CircularProgress } from "@mui/material";
 import { Box, Button, InputAdornment, TextField } from "@mui/material";
 
-import { useMultiMQTTRoomContext } from "@services/MQTTRoomService/react";
+import { useMultiMQTTRoomContext } from "@services/MultiMQTTRoomService/react";
 
 import customLogger from "@utils/customLogger";
