--- src/services/MQTTRoomService/index.ts
@@ -1,6 +1,0 @@
-import MQTTRoom from "./MQTTRoom";
-import type { MQTTRoomEvents } from "./MQTTRoom.sharedBindings";
-import validateTopic from "./validateTopic";
-
-export { MQTTRoom, validateTopic };
-export type { MQTTRoomEvents };
