+++ src/services/MultiMQTTRoomService/MQTTRoom/MQTTRoom.utils.ts
