+++ src/services/MultiMQTTRoomService/MQTTRoom/index.ts
@@ -1,0 +1,3 @@
+import MQTTRoom from "./MQTTRoom";
+
+export default MQTTRoom;
