+++ src/services/MultiMQTTRoomService/index.ts
@@ -1,0 +1,8 @@
+import MQTTRoom from "./MQTTRoom/MQTTRoom";
+import type { MQTTRoomEvents } from "./MQTTRoom/MQTTRoom.sharedBindings";
+import MultiMQTTRoomService from "./MultiMQTTRoomService";
+import validateTopic from "./validateTopic";
+
+export default MultiMQTTRoomService;
+export { MQTTRoom, validateTopic };
+export type { MQTTRoomEvents };
